Split utils.py into utils.py and _utils.py
