+++ drf_logger/mixins.py
@@ -1,7 +1,6 @@
 from django.utils import timezone
 from drf_logger import utils
-
-LOG_LEVELS = ('DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL')
+from drf_logger._utils import _get_client_ip
 
 
 class APILoggingMixin:
@@ -11,7 +10,7 @@ class APILoggingMixin:
     def dispatch(self, request, *args, **kwargs):
         extra: dict = {}
         extra['time']: str = str(timezone.now())
-        extra['ip']: str = utils.get_client_ip(request)
+        extra['ip']: str = _get_client_ip(request)
         extra['user_id'] = request.user.id
         extra['method']: str = request.method
